LPS-112066 Extract specific Index logic to a new class
